+++ contracts/token/commands/commands.ts
@@ -1,0 +1,4 @@
+export { deploy } from "./deploy";
+export { mint } from "./mint";
+export { transfer } from "./transfer";
+export { transferAndCall } from "./transferAndCall";
